Added support for reuse distance.
The new -bf-reuse-dist option tells Byfl to keep track of data reuse
distance, defined as the median number of unique memory addresses
referenced between successive references to the same address.  This
gives some indication of a program's data locality.

This version of the code was tested with Subversion revision 165745 of
LLVM/Clang/DragonEgg.