Fixed memory-freeing bug
Due to a shallow copy in the code, the inst_mix_histo *pointer* was
getting copied from one object to another, then double-freed.  This
version allocates inst_mix_histo as a fixed-size array both to solve
that problem and to reduce the number of memory allocations (at the
expense of storage requirements when -bf-inst-mix isn't used).