Altered calculation of ops and op bits
This revision attempts to rationalize the calculation of ops and op
bits.  All operations except for a few (phi nodes, debug intrinsics,
landing pads, and bitcasts) are now considered "ops".  Op bits are
computed on all operations except loads and stores, the goal being to
maintain A=B+C and A=-B as having a bit:op-bit ratio of 1, regardless
of the types of A, B, and C.