Post-Byfl optimizations
Modified the wrapper scripts to run "opt -O3" on the bitcode after
Byfl instrumentation, even if the pre-Byfl code was optimized at -O2
or lower.  This should make the instrumented code run faster without
affecting the results.  The only cost is in additional compilation
time, which I hope is not a critical issue.