Added a bf-inst script
bf-inst is a wrapper script for LLVM's opt command that knows how to
apply Byfl instrumentation to an LLVM bitcode file (binary or ASCII)
or to the bitcode contents of an ar archive file and then re-run opt
to optimize the instrumented code.  It can also be used to output
command-line options for Clang to instruct it to apply a Byfl pass as
part of its code-transformation activities.