Renamed -bf-byte-counts to -bf-mem-footprint
Replace the generic-sounding and hard-to-remember -bf-byte-counts
command-line option with -bf-mem-footprint, which, while not perfectly
accurate, is at least likely to be a bit more mnemonic to users.
Updated the README file to describe -bf-mem-footprint (including in
the sample Byfl output) and removed all mention of -bf-reuse-dist, as
reuse-distance calculations are now deprecated for being unbearably
slow and memory-hungry.  Byfl continues to compile and link against
the reuse-distance code, though, just in case we ever need to do a
performance comparison between -bf-mem-footprint and -bf-reuse-dist.