Minimum bash version
Because build-llvm-byfl uses bash's coproc statement, the script
doesn't work with versions of bash prior to 4.0 as these lack coproc
support.  Alas, Apple ships a 3.x version of bash with the current
OS X, which means the script can't run on a "pure" OS X
installation.  As a stopgap measure, the script now explicitly
checks BASH_VERSION and aborts with a more user-friendly message
than "syntax error", which is what would happen otherwise.