Handle multiply specified input files
This revision of the wrapper scripts correctly handles the obscure
special case of the same input file being specified multiple times on
the same command line.  In this case, the .s file is deleted/moved
after the first time the input file is processed, which would cause
opt to fail to find its input file and therefore abort.