Disable -Wcovered-switch-default
If Byfl is configured using clang++ as the C++ compiler, the configure
script detects that -Wcovered-switch-default is supported and includes
that in CXXFLAGS.  Unfortunately, some parts of the Byfl build require
g++, which does not recognize this flag and consequently aborts when
passed it.  In this revision, -Wcovered-switch-default is never used,
even if it's detected by configure.

This problem was reported by Christoph Junghans.