Fixed multiple-Perl-version bug
The installation script used to run into problems if the Byfl build
directory contained remnants from prior builds created using different
Perl versions.  This situation caused the byflperllib variable to
contain an invalid filename, which broke the test of bf-gcc.

The solution represented by this revision is to add a line to
wrappers.mak to output the contents of the INSTALLSITELIB make
variable.  The installation script then uses that variable to set
byflperllib.