Dynamically resize cache to account for all accesses.
Rather than specifying the absolute size of the cache, allow it to grow
in size dynamically. When a line misses in the cache, increase the cache
size by a single line and set its hit count to the LRU slot's hit count
plus one. This also removes the need to maintain an access count per
line, as that would only give the number of times the nth line of an n
line cache would be accesses.