One one access happens per line, not per byte.
Before, each byte address requested would count as a separate request.
This has been replaced with a count of the number of cache lines
touched. It's possible that all addresses fit entirely inside a single
line, and therefore only count as a single access. However, it's
possible that the request is not aligned, and would then require
multiple cache requests.