Simplify handling of cold misses.
The cold miss counter is equivalent to the number of unique addresses in
the cache. This in turn is equivalent to the number of lines in the
cache, and therefor doesn't need to be recorded. As well, any time there
is a miss in the cache, don't record it as a hit; this communicates all
cold misses the same way as the counter did.