Add flag to control maximum number of cache sets.
Byfl compile flag -bf-max-set-bits allows the user to put a limit on the
number of cache sets to be modeled. This is a practical limit: it by
definition has to be smaller than size of the address space, but due to
implementation an unbounded limit would attempt to allocate the entire
memory space.