Build script accepts new -C and -D options
The -C option is like -c (compile only) but additionally suppresses
the LLVM and Byfl configuration steps.  This is convenient because it
enables LLVM to rebuild only modified files without having to rebuild
all of LLVM, which can take a very long time.  The -D option is how
the script provides an interface to the DESTDIR build variable.  It
enables staging the installation into a temporary directory, which can
be useful if the user lacks write access to the final installation
directory.