Begin complete reworking of the build process
Byfl previously inherited a lot of cruft from LLVM's build process.
Some of that is because Byfl was based on an in-source example, but
this adds an artificial dependency on an LLVM source tree.  This
revision is the beginning of an attempt to throw all that away and
start fresh with a proper autoconf/automake/libtool build process.