Replaced AC_CHECK_FILE with "test -f"
Autoconf's AC_CHECK_FILE is largely useless because it purportedly
checks for file existence not on the build machine but on the host
machine.  (And how does it know what host machine I'll use and how to
log onto it?  Beats me.)  While this is in fact desirable for the
check for dragonegg.so, the problem is that AC_CHECK_FILE aborts
configure if the file is not found during a cross-compile, while I'd
prefer just to issue a warning message.  Replacing AC_CHECK_FILE with
"test -f" enables the configure script to issue a warning even when
cross-compiling.