Moved binarytagdefs.h and byfl-common.h to top-level include
Because binarytagdefs.h is used by the library in lib/byfl and will
soon be used by programs in the tools/postproc directory, it makes
sense to hoist that above both lib/byfl and tools/postproc (i.e., to
the top level).  While I was at it, I moved byfl-common.h from
lib/include to the top-level include directory because I've always
found it a bit goofy to have a include directory beneath a lib
directory.