Logged more basic-block information in binary format
Currently, the only fields in ByteFlopCounters that aren't written to
the binary output file are the mem_insts and inst_mix_histo arrays.
These are large -- totaling ~200 fields -- and of questionable use on
the basic-block level relative to their size.  Perhaps they can be
included if some command-line option is specified.