Removed bfbin2netcdf
bfbin2netcdf reached a dead end because it didn't seem like there was
a way to generate a table of mixed datatypes that could be viewed with
hdfview.  bfbin2hdf5, which was introduced in the previous commit, can
parse as much of a binary Byfl files as bfbin2netcdf could and can
additionally produce mixed-datatype tables that appear correctly in
hdfview.