Changed the way the PRNG is seeded
Previously, only the module identifier (typically, the name of the
source file) was being used to seed the pseudo random number
generator.  In this revision, the seed is based on a function of the
module identifier, the process ID, and the current time in
nanoseconds.  That ought to make seed conflicts extremely rare, even
with parallel builds.

This commit closes #8 (I hope).