Fixed integer overflow in -bf-mem-footprint output
As reported by Chris Daley (@cdaley), the 100% memory-footprint data
is inconsistent with the total number of unique bytes due to the tally
of access counts overflowing.  Chris's patch (reported in #9) was
simply to promote the tally to a 64-bit integer while still tracking
access counts with 32-bit integers (to reduce the footprint of Byfl's
bookkeeping data).

This commit closes #9.