New sort order for memory footprint
Previously, the "X bytes cover Y% of memory accesses" histogram sorted
in decreasing order of tally.  For example, if 3 addresses were each
accessed 10 times and 4 addresses were each accessed 8 times, the
former would appear first in the sorted order.  This revision sorts by
the product of the number of addresses and the number of accesses.
Hence, if, as before, 3 addresses were each accessed 10 times and 4
addresses were each accessed 8 times, the latter would appear first in
the sorted order because 3*10=30 < 4*8=32.