New -x option to build-llvm-byfl
The build-llvm-byfl do-everything installation script now honors a -x
option that specifies a component *not* to download or build.
Supported values are "llvm" (which includes both LLVM and Clang),
"dragonegg", and "byfl".  The -x option is expected to be useful on
systems that don't support DragonEgg and on systems that require
manual steps to build one or more Byfl components.