Removed redundant script copies
Previously, tools/wrappers/Makefile.am was copying bf-gcc to bf-g++,
bf-gfortran, and bf-gccgo.  However, tools/wrappers/Makefile.PL
already specifies that bf-gcc should be symlinked to each of those
other scripts.  This led to a race condition during parallel builds
that could result in a "cp: ‘bf-gcc’ and ‘bf-gccgo’ are the same file"
error.  In this new revision, which ought to close issue #10, the
copies are removed and only the symlinks are retained.