Bug fix: Track allocations immediately, not at the end of the basic block
Byfl was observing a lot of unknown data structures.  The problem
appears to have been that one basic block would invoke malloc() then
call a function that uses the allocated memory.  Because Byfl inserts
the malloc-tracking code at the *end* of the basic block, the
allocation is not being tracked at the time the child function is
called.