Instrument data-structure accesses immediately
Because malloc() and free() have to be instrumented immediately in
order for function calls to have access to the current
memory-allocation state, all data-structure accesses also have to be
instrumented immediately.  Otherwise, the instrumentation of a load or
store may follow the instrumentation of a free() of the same data.