Don't install GCC wrapper scripts if DragonEgg is not available
DragonEgg development seems to have halted circa November 2014.
DragonEgg fails to build against either GCC 4.9+ or LLVM 3.6+ (despite
there being a DragonEgg 3.6 release).  The last working combination
appears to be revision 223694 of LLVM, Clang, and DragonEgg and
version 4.8 of GCC, and that's what this revision of Byfl was tested
against.