Updated LLVM 3.6 APIs to LLVM trunk (3.7) APIs
Module::getDataLayout() now returns a pointer instead of a value.
ConstantExpr::getGetElementPtr() and GetElementPtrInst::Create() now
take an extra argument, an LLVM Type*.  We no longer invoke
getAnalysisUsage() because DataLayoutPass and LoopInfo were giving
error messages, and the code seems to work without those.

This revision ought to be audited to ensure I haven't broken
correctness in my attempt to get the code to compile, link, and run.

:warning: The changes made for this revision break compatibility with
Byfl's llvm-3.6 (and, of course, llvm-3.5) branch!