Added -bf-inst-deps for instruction dependencies
This new feature reports what instructions produced the non-constant
operands for each other instruction.  It may be used to identify
candidates for fused instructions (like fused multiply-add) or simply
instruction pairs that need to have low latency/high throughput in the
instruction pipeline.  It may also be used to produce a synthetic code
with the same instruction mix and inter-instruction dependencies as a
real code.