Removed all dependencies on both backtrace() and the BFD library
For per-basic-block and per-data-structure accesses, the Byfl run-time
library previously identified callee source-code location at run time.
It called the glibc backtrace() function to walk the stack then
various functions from the BFD library to map the most deeply nested
address lying within the user's application to a source-code location.
There are at least four problems with this approach:

  1) Distinguishing application program counters from Byfl run-time
     library program counters is based on a heuristic.  While I've not
     yet observed it fail in practice, there's certainly no guarantee
     that the heuristic will work in all cases.

  2) The BFD developer files are seldom available on stock OS
     installations.  Users must therefore build libbfd from source,
     making sure to target the version that exactly matches the GNU
     Binutils libraries and binaries installed on the system.  Failing
     to do so prevents Byfl from reporting source-code locations,
     which greatly reduces the tool's utility.

  3) backtrace() and the BFD library incur a run-time cost (walking
     the call stack and the executable's symbol table) to infer
     information that is already known at compile time.

  4) Richer symbol information is available at compile time than at
     run time.  For example, the source-code location at which a
     global variable was declared does not seem to be available at run
     time.

The current commit restructures the Byfl plugin to read at compile
time the LLVM debug metadata associated with LLVM IR statements and
data structures, extract source-code locations, and pass this
information to the Byfl run-time library to log.

This commit represents a sizable change to the Byfl source code: a
total of 17 files and thousands of lines of code were modified,
deleted, or added.

[This is a cherry-pick commit from the llvm-3.5 branch, modified to
work with the latest LLVM APIs.]

Conflicts:
	lib/bytesflops/helpers.cpp