Worked around LLVM's pathetically slow Value::print() method
Depending on function length, Value::print() can take upwards of a
second per invocation.  (It appears to be constructing some sort of
mapping of all Values in a function on every invocation with no
caching of prior results.)  This revision incorporates a grotesque
hack in which, at function-initialization time, the entire function is
converted to a single string then parsed explicitly into
per-instruction strings for later use.  I've measured this approach as
being ~30x faster (when compiling and instrumenting NuT) than calling
Value::print() on individual instructions.  The problem is that this
hack will cease to work if LLVM ever changes the way that ASCII
bitcode is formatted.

This version of Byfl was tested with LLVM/Clang revision 241287.