Replace bf2hpctk with bfbin2hpctk
The old bf2hpctk Perl script was not particular robust in that it
needed to parse the program's textual output.  The new bfbin2hpctk
program instead processes the structured, binary .byfl file.  It
extracts data from the Functions table and uses that to produce a
complete HPCToolkit database directory, not just the constituent
experiment.xml file.