Replace bf2cgrind with bfbin2cgrind
The old bf2cgrind Perl script was not particular robust in that it
needed to parse the program's textual output.  The new bfbin2cgrind
program instead processes the structured, binary .byfl file.  It
extracts data from the Functions table (and some metadata from a few
other tables) and uses that to produce a Callgrind file.