Improved Travis CI support
Travis CI's build systems are running a version of Automake that
predates the introduction of AM_TESTS_ENVIRONMENT, upon which the Byfl
test harness relies.  This revision works around that limitation by
setting TESTS_ENVIRONMENT, which is supported, to AM_TESTS_ENVIRONMENT
unless it already has a value.

Also, -lstdc++ is once again added the the link line, at least for
bf-g++.  I encountered a system on which that was required.