Workaround for failed instruction stringification
Before LLVM 3.9, Byfl never failed to convert Instructions to strings.
As of LLVM revision 267798 (but perhaps a bit earlier), many Br
instructions are failing to convert.  Until I have time to track down
the source of the problem, I added a workaround that simply falls back
to a much slower mechanism when necessary.