Prevent LLVM from unloading the Byfl plugin until exit
As discussed in
https://stackoverflow.com/questions/51146196/what-happens-to-the-thread-spawned-by-a-shared-library-upon-dlclose,
unloading a plugin can be fraught with problems.  That Stack Overflow
discussion recommends leaking a reference to the plugin to prevent it
from being unloaded.  This means we no longer need to explicitly pass
-Wl,-znodelete to the linker, which we don't want to do because that
appears to be a platform-specific linker feature (e.g., Linux but not
OS X).