Localize the HDF5 compiler flags and libraries to bfbin2hdf5
Previously, we were trying (unsuccessfully) to use the h5c++ wrapper
script to build everything in the tools/postproc directory.  The two
problems caused by that approach are that (1) only bfbin2hdf5 ought to
be built with h5c++, not all of the postprocessing tools, and (2)
h5c++ seems extremely flaky.  It often fails to parse correctly the
C++ compiler's command-line arguments.

In this revision, we replace h5c++ with target-specific definitions,
include directories, and libraries using the values that FindHDF5
provides.