Work around pkg-config stupidity
pkg-config automatically removes duplicate flags from Cflags and Libs,
and there's apparently no way to disable this feature.  Hence, it
converts `-Xclang -load -Xclang bytesflops.so` to
`-Xclang -load bytesflops.so`, which leads to a complaint from the
compiler.  As a workaround, we now specify `-fplugin=bytesflops.so`.
This has the side effect of always running the plugin, but that's
typically what a user would want to do, anyway.